Issue #1: add support of default attribute values
WIP: added Value to Attribute and regenerated code
So in Model Editor (FM) the value can be defined and used in own code
generator as fallback if not AttributeValue was defined by user in
Feature Configuration (FC).

TODO:
- show the default value in FC
- add something like "use default value" checkbox in FC, if checked copy
the value from FM to FC and store that is was copied not entered. If not
checked then delete AttributeValue if was copied from FM before or just
empty the value inside AttributeValue... just thinking...

https://bitbucket.org/jpikl/yafmt/issues/1/add-support-of-default-attribute-values